adding test for GW account validation
