+++ handlers/new-product-api/src/test/scala/com/gu/newproduct/api/addsubscription/validation/voucher/PaperAccountValidationTest.scala
@@ -1,26 +1,26 @@
 package com.gu.newproduct.api.addsubscription.validation.voucher
 
-import com.gu.i18n.Currency.{GBP, USD}
-import com.gu.newproduct.api.addsubscription.validation.paper.PaperAccountValidation
+import com.gu.i18n.Currency.GBP
+import com.gu.newproduct.api.addsubscription.validation.guardianweekly.GuardianWeeklyAccountValidation
 import com.gu.newproduct.api.addsubscription.validation.{Failed, Passed, ValidatedAccount}
-import com.gu.newproduct.api.addsubscription.zuora.GetAccount.{AccountBalanceMinorUnits, AutoPay, PaymentMethodId, SfContactId}
+import com.gu.newproduct.api.addsubscription.zuora.GetAccount.{AccountBalanceMinorUnits, AutoPay, IdentityId, PaymentMethodId}
 import org.scalatest.{FlatSpec, Matchers}
 
 class PaperAccountValidationTest extends FlatSpec with Matchers {
   val account = ValidatedAccount(
-    identityId = None,
-    sfContactId = Some(SfContactId("sfContactId")),
+    identityId = Some(IdentityId("1234")),
+    sfContactId = None,
     paymentMethodId = PaymentMethodId("id"),
     autoPay = AutoPay(true),
     accountBalanceMinorUnits = AccountBalanceMinorUnits(1233),
     currency = GBP
   )
-  it should "pass if currency in account is GBP " in {
-    PaperAccountValidation(account) shouldBe Passed(account)
+  it should "pass if account is valid" in {
+    GuardianWeeklyAccountValidation(account) shouldBe Passed(account)
   }
 
-  it should "fail if currency in account is not GBP " in {
-    val dollarAccount = account.copy(currency = USD)
-    PaperAccountValidation(dollarAccount) shouldBe Failed("Invalid currency in Zuora account: USD. Only GBP is allowed for the selected plan")
+  it should "fail if identity id is not present" in {
+    val noIdentityIdAccount = account.copy(identityId = None)
+    GuardianWeeklyAccountValidation(noIdentityIdAccount) shouldBe Failed("Zuora account has no Identity Id")
   }
 }
